requirements.txt
click
python-json-logger
pyzmq